add split and part1

+++ lib/ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: rkina <rkina@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/07 17:55:50 by rkina             #+#    #+#             */
-/*   Updated: 2019/09/10 19:21:59 by rkina            ###   ########.fr       */
+/*   Updated: 2019/09/12 15:55:17 by rkina            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_strlen(const char *s)
+#include "libft.h"
+
+size_t		ft_strlen(const char *s)
 {
-	int		len;
+	size_t		len;
 
 	len = 0;
 	while (s[len])
